+++ lib/qonfig/validation.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# @api private
+# @since 0.19.0
+module Qonfig::Validation
+  require_relative 'validation/validators'
+  require_relative 'validation/collections'
+  require_relative 'validation/building'
+  require_relative 'validation/predefinition_mixin'
+end