+++ lib/qonfig/validation/building/instance_builder.rb
@@ -1,32 +1,32 @@
 # frozen_string_literal: true
 
 # @api private
-# @since 0.13.0
-class Qonfig::Validator::Builder
-  require_relative 'builder/attribute_consistency'
+# @since 0.20.0
+class Qonfig::Validation::Building::InstanceBuilder
+  require_relative 'instance_builder/attribute_consistency'
 
   # @return [NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   EMPTY_SETTING_KEY_PATTERN = nil
 
   # @return [NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   NO_RUNTIME_VALIDATION_METHOD = nil
 
   # @return [NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   NO_VALIDATION_LOGIC = nil
 
   # @return [NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   NO_PREDEFINED_VALIDATOR = nil
 
   # @return [Boolean]
@@ -36,6 +36,7 @@ class Qonfig::Validator::Builder
   DEFAULT_STRICT_BEHAVIOUR = false
 
   class << self
+    # @param data_set_klass [Class<Qonfig::DataSet>]
     # @option setting_key_pattern [String, Symbol, NilClass]
     # @option predefined_validator [String, Symbol, NilClass]
     # @option runtime_validation_method [String, Symbol, NilClass]
@@ -44,8 +45,9 @@ class Qonfig::Validator::Builder
     # @return [Qonfig::Validator::MethodBased, Qonfig::Validator::ProcBased]
     #
     # @api private
-    # @since 0.13.0
+    # @since 0.20.0
     def build(
+      data_set_klass,
       setting_key_pattern: EMPTY_SETTING_KEY_PATTERN,
       runtime_validation_method: NO_RUNTIME_VALIDATION_METHOD,
       validation_logic: NO_VALIDATION_LOGIC,
@@ -53,6 +55,7 @@ class Qonfig::Validator::Builder
       predefined_validator: NO_PREDEFINED_VALIDATOR
     )
       new(
+        data_set_klass,
         setting_key_pattern,
         predefined_validator,
         runtime_validation_method,
@@ -62,6 +65,7 @@ class Qonfig::Validator::Builder
     end
   end
 
+  # @param data_set_klass [Class<Qonfig::DataSet>]
   # @param setting_key_pattern [String, Symbol, NilClass]
   # @param predefined_validator_name [String, Symbol, NilClass]
   # @param runtime_validation_method [String, Symbol, NilClass]
@@ -70,14 +74,16 @@ class Qonfig::Validator::Builder
   # @return [void]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def initialize(
+    data_set_klass,
     setting_key_pattern,
     predefined_validator_name,
     runtime_validation_method,
     strict,
     validation_logic
   )
+    @data_set_klass = data_set_klass
     @setting_key_pattern = setting_key_pattern
     @predefined_validator_name = predefined_validator_name
     @runtime_validation_method = runtime_validation_method
@@ -85,10 +91,12 @@ class Qonfig::Validator::Builder
     @validation_logic = validation_logic
   end
 
-  # @return [Qonfig::Validator::MethodBased, Qonfig::Validator::ProcBased]
+  # @return [Qonfig::Validation::Validators::MethodBased]
+  # @return [Qonfig::Validation::Validators::ProcBased]
+  # @return [Qonfig::Validation::Validators::Predefined]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build
     validate_attributes!
     build_validator
@@ -96,22 +104,28 @@ class Qonfig::Validator::Builder
 
   private
 
+  # @return [Class<Qonfig::DataSet>]
+  #
+  # @api private
+  # @since 0.20.0
+  attr_reader :data_set_klass
+
   # @return [String, Symbol, NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :setting_key_pattern
 
   # @return [String, Symbol, NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :predefined_validator_name
 
   # @return [String, Symbol, NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :runtime_validation_method
 
   # @return [Boolean]
@@ -123,7 +137,7 @@ class Qonfig::Validator::Builder
   # @return [Proc, NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :validation_logic
 
   # @return [void]
@@ -131,9 +145,9 @@ class Qonfig::Validator::Builder
   # @raise [Qonfig::ArgumentError]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def validate_attributes!
-    AttributeConsistency.check!(
+    AttributesConsistency.check!(
       setting_key_pattern,
       predefined_validator_name,
       runtime_validation_method,
@@ -142,10 +156,12 @@ class Qonfig::Validator::Builder
     )
   end
 
-  # @return [Qonfig::Validator::MethodBased, Qonfig::Validator::PorcBased]
+  # @return [Qonfig::Validation::Validators::MethodBased]
+  # @return [Qonfig::Validation::Validators::ProcBased]
+  # @return [Qonfig::Validation::Validators::Predefined]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build_validator
     case
     when predefined_validator_name then build_predefined_validator
@@ -157,38 +173,47 @@ class Qonfig::Validator::Builder
   # @return [Qonfig::Settings::KeyMatcher, NilClass]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build_setting_key_matcher
     Qonfig::Settings::KeyMatcher.new(setting_key_pattern.to_s) if setting_key_pattern
   end
 
-  # @return [Qonfig::Validator::MethodBased]
+  # @return [Qonfig::Validation::Validators::MethodBased]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build_method_based_validator
-    Qonfig::Validator::MethodBased.new(
+    Qonfig::Validation::Validators::MethodBased.new(
       build_setting_key_matcher, strict, runtime_validation_method
     )
   end
 
-  # @return [Qonfig::Validator::ProcBased]
+  # @return [Qonfig::Validation::Validators::ProcBased]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build_proc_based_validator
-    Qonfig::Validator::ProcBased.new(
+    Qonfig::Validation::Validators::ProcBased.new(
       build_setting_key_matcher, strict, validation_logic
     )
   end
 
-  # @return [Qonfig::Settings::Predefined]
+  # @return [Qonfig::Validation::Validators::Predefined]
+  #
+  # @see Qonfig::Validation::Collections::PredefinedRegistry
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def build_predefined_validator
-    Qonfig::Validator::Predefined.build(
-      predefined_validator_name, build_setting_key_matcher, strict
+    predefined_validation_logic =
+      begin
+        data_set_klass.predefined_validators.resolve(predefined_validator_name)
+      rescue Qonfig::ValidatorNotFoundError
+        Qonfig::DataSet.predefined_validators.resolve(predefined_validator_name)
+      end
+
+    Qonfig::Validation::Validators::Predefined.new(
+      build_setting_key_matcher, strict, predefined_validation_logic
     )
   end
 end