+++ lib/qonfig/validation/collections.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# @api private
+# @since 0.20.0
+module Qonfig::Validation::Collections
+  require_relative 'collections/predefined_registry'
+  require_relative 'collections/instance_collection'
+end