+++ lib/qonfig/validation/collections/instance_collection.rb
@@ -1,32 +1,32 @@
 # frozen_string_literal: true
 
 # @api private
-# @since 0.13.0
-class Qonfig::Validator::Collection
+# @since 0.20.0
+class Qonfig::Validation::Collections::InstanceCollection
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   include Enumerable
 
-  # @return [Array<Qonfig::Validator::MethodBased,Qonfig::Validator::ProcBased>]
+  # @return [Array<Qonfig::Validation::Validators::Base>]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :validators
 
   # @return [void]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def initialize
     @validators = []
     @access_lock = Mutex.new
   end
 
-  # @param validator [Qonfig::Validator::MethodBased, Qonfig::Validator::ProcBased]
+  # @param validator [Qonfig::Validation::Validators::Base]
   # @return [void]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def add_validator(validator)
     thread_safe { validators << validator }
   end
@@ -36,27 +36,29 @@ class Qonfig::Validator::Collection
   # @return [Enumerable]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def each(&block)
     thread_safe { block_given? ? validators.each(&block) : validators.each }
   end
 
-  # @param collection [Qonfig::Validator::Collection]
+  # @param collection [Qonfig::Validation::Collections::InstanceCollection]
   # @return [void]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def concat(collection)
     thread_safe { validators.concat(collection.validators) }
   end
 
-  # @return [Qonfig::Validator::Collection]
+  # @return [Qonfig::Validation::Collections::InstanceCollection]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def dup
     thread_safe do
-      self.class.new.tap { |duplicate| duplicate.concat(self) }
+      Qonfig::Validation::Collections::InstanceCollection.new.tap do |duplicate|
+        duplicate.concat(self)
+      end
     end
   end