+++ lib/qonfig/validation/validators.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+# @api private
+# @since 0.20.0
+module Qonfig::Validation::Validators
+  require_relative 'validators/basic'
+  require_relative 'validators/method_based'
+  require_relative 'validators/proc_based'
+  require_relative 'validators/predefined'
+  require_relative 'validators/composite'
+end