+++ lib/qonfig/validation/validators/proc_based.rb
@@ -1,33 +1,33 @@
 # frozen_string_literal: true
 
 # @api private
-# @since 0.13.0
-class Qonfig::Validator::Predefined::Common < Qonfig::Validator::Basic
+# @since 0.20.0
+class Qonfig::Validation::Validators::ProcBased < Qonfig::Validation::Validators::Basic
   # @return [Proc]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   attr_reader :validation
 
-  # @param setting_key_matcher [Qonfig::Settings::KeyMatcher]
+  # @param setting_key_matcher [Qonfig::Settings::KeyMatcher, NilClass]
   # @param strict [Boolean]
-  # @param validation [Proc]
+  # @param vaidation [Proc]
   # @return [void]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def initialize(setting_key_matcher, strict, validation)
     super(setting_key_matcher, strict)
     @validation = validation
   end
 
   # @param data_set [Qonfig::DataSet]
-  # @return [void]
+  # @return [Boolean]
   #
   # @raise [Qonfig::ValidationError]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def validate_concrete(data_set)
     data_set.settings.__deep_each_setting__ do |setting_key, setting_value|
       next unless setting_key_matcher.match?(setting_key)
@@ -36,20 +36,20 @@ class Qonfig::Validator::Predefined::Common < Qonfig::Validator::Basic
       raise(
         Qonfig::ValidationError,
         "Invalid value of setting <#{setting_key}> (#{setting_value})"
-      ) unless validation.call(setting_value)
+      ) unless data_set.instance_exec(setting_value, &validation)
     end
   end
 
   # @param data_set [Qonfig::DataSet]
-  # @return [void]
+  # @return [Boolean]
   #
   # @raise [Qonfig::ValidationError]
   #
   # @api private
-  # @since 0.13.0
+  # @since 0.20.0
   def validate_full(data_set)
-    # :nocov:
-    raise Qonfig::Error, 'Predefined validator should be used only with a setting key'
-    # :nocov:
+    unless data_set.instance_eval(&validation)
+      raise(Qonfig::ValidationError, 'Invalid config object')
+    end
   end
 end